images/utils/base/requirements.txt
filelock
pyyaml
